Refactoring bootstrap
