+++ src/test/kotlin/no/nav/syfo/service/SoknadServiceSpek.kt
@@ -119,8 +119,6 @@ object SoknadServiceSpek : Spek( {
                 testDatabase.connection.erSoknadLagret(soknadRecord) shouldBe true
                 testDatabase.hentSoknadsData(LocalDateTime.of(2019,8,2,0,0),
                         LocalDateTime.of(2019,8,3,0,0))[0].antall shouldEqual 1
-//                testDatabase.hentSoknadsData(LocalDateTime.of(2019,8,3,0,0),
-//                        LocalDateTime.of(2019,8,4,0,0)) shouldEqual emptyList()
         }
 
         it ("søknad kan lagres i loggtabell") {
